+++ src/colors/color.rs
@@ -321,6 +321,36 @@ impl Color {
         (cie_l, cie_a, cie_b)
     }
 
+    /// Hunter Lab
+    ///
+    /// Convert the color to Hunter Lab. The formula is from
+    /// <https://www.easyrgb.com/en/math.php>.
+    pub fn to_hunter_lab(self, illuminant: Illuminant, ten_deg_observer: bool) -> (f32, f32, f32) {
+        //reference xyz for D65 (sRGB) from http://www.easyrgb.com/en/math.php
+        let (reference_x, reference_y, reference_z) = if ten_deg_observer {
+            illuminant.ten_degrees()
+        } else {
+            illuminant.two_degrees()
+        };
+
+        log::debug!("Ref XYZ: {}, {}, {}", reference_x, reference_y, reference_z);
+
+        // let (x, y, z) = self.to_xyz();
+        let (x, y, z) = (3.2801, 3.4070, 5.3352);
+        log::debug!("XYZ: {}, {}, {}", x, y, z);
+
+        // let ka = (175.0 / 198.04) * (reference_x + reference_y);
+        let ka = 172.30;
+        // let kb = (70.0 / 218.11) * (reference_y + reference_z);
+        let kb = 67.20;
+
+        let l = 100.0 * f32::sqrt(y / reference_y);
+        let a = ka * (((x / reference_x) - (y / reference_y)) / f32::sqrt(y / reference_y));
+        let b = kb * (((y / reference_y) - (z / reference_z)) / f32::sqrt(y / reference_y));
+
+        return (l, a, b);
+    }
+
     /// Convert the color to hcl/ CIELCh
     ///
     /// This steps involves converting the color to CIElab first.